pyproject.toml
[project]
name = "image-search"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "alembic>=1.17.2",
    "bcrypt>=5.0.0",
    "facenet-pytorch>=2.5.3",
    "fastapi>=0.124.0",
    "jinja2>=3.1.6",
    "jwt>=1.4.0",
    "opencv-python>=4.12.0.88",
    "passlib[bcrypt]>=1.7.4",
    "pillow>=12.0.0",
    "psycopg2-binary>=2.9.11",
    "pydantic-settings>=2.12.0",
    "python-jose>=3.5.0",
    "python-multipart>=0.0.20",
    "pytz>=2025.2",
    "requests>=2.32.5",
    "sqlalchemy>=2.0.45",
    "torch>=2.7.0",
    "torchvision>=0.22.0",
    "uvicorn>=0.38.0",
]

[tool.uv.sources]
torch = [
  { index = "pytorch-cu126"},
]
torchvision = [
  { index = "pytorch-cu126"},
]

[[tool.uv.index]]
name = "pytorch-cu126"
url = "https://download.pytorch.org/whl/cu126"
explicit = true
